python/requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-cloud-texttospeech==0.5.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
oauthlib==3.1.0
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyDispatcher==2.0.5
pyglet==1.4.8
pyserial==3.4
python-openzwave==0.4.19
python-twitter==3.5
pyttsx3==2.71
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
